+++ lockfile/src/lib.rs
@@ -4,7 +4,7 @@ use std::path::{Path, PathBuf};
 use std::str::FromStr;
 
 pub use cargo::Cargo;
-pub use csharp::{CSProj, PackagesLock};
+pub use csharp::PackagesLock;
 pub use golang::GoSum;
 use ignore::WalkBuilder;
 pub use java::{GradleLock, Pom};
@@ -48,11 +48,8 @@ pub enum LockfileFormat {
     #[serde(alias = "maven")]
     Maven,
     Gradle,
-    // This is historically called "nuget" but it's actually for MSBuild project files.
-    // Nuget has its own file formats that are not currently supported.
-    #[serde(alias = "nuget")]
-    Msbuild,
-    NugetLock,
+    #[serde(alias = "nugetlock")]
+    Nuget,
     Go,
     Cargo,
     Spdx,
@@ -91,8 +88,7 @@ impl LockfileFormat {
             LockfileFormat::Poetry => "poetry",
             LockfileFormat::Maven => "mvn",
             LockfileFormat::Gradle => "gradle",
-            LockfileFormat::Msbuild => "msbuild",
-            LockfileFormat::NugetLock => "nugetlock",
+            LockfileFormat::Nuget => "nuget",
             LockfileFormat::Go => "go",
             LockfileFormat::Cargo => "cargo",
             LockfileFormat::Spdx => "spdx",
@@ -111,8 +107,7 @@ impl LockfileFormat {
             LockfileFormat::Poetry => &Poetry,
             LockfileFormat::Maven => &Pom,
             LockfileFormat::Gradle => &GradleLock,
-            LockfileFormat::Msbuild => &CSProj,
-            LockfileFormat::NugetLock => &PackagesLock,
+            LockfileFormat::Nuget => &PackagesLock,
             LockfileFormat::Go => &GoSum,
             LockfileFormat::Cargo => &Cargo,
             LockfileFormat::Spdx => &Spdx,
@@ -147,11 +142,10 @@ impl Iterator for LockfileFormatIter {
             6 => LockfileFormat::Pipenv,
             7 => LockfileFormat::Maven,
             8 => LockfileFormat::Gradle,
-            9 => LockfileFormat::NugetLock,
-            10 => LockfileFormat::Msbuild,
-            11 => LockfileFormat::Go,
-            12 => LockfileFormat::Cargo,
-            13 => LockfileFormat::Spdx,
+            9 => LockfileFormat::Nuget,
+            10 => LockfileFormat::Go,
+            11 => LockfileFormat::Cargo,
+            12 => LockfileFormat::Spdx,
             _ => return None,
         };
         self.0 += 1;
@@ -342,8 +336,7 @@ mod tests {
             ("package-lock.json", LockfileFormat::Npm),
             ("npm-shrinkwrap.json", LockfileFormat::Npm),
             ("pnpm-lock.yaml", LockfileFormat::Pnpm),
-            ("sample.csproj", LockfileFormat::Msbuild),
-            ("packages.lock.json", LockfileFormat::NugetLock),
+            ("packages.lock.json", LockfileFormat::Nuget),
             ("gradle.lockfile", LockfileFormat::Gradle),
             ("effective-pom.xml", LockfileFormat::Maven),
             ("requirements.txt", LockfileFormat::Pip),
@@ -374,9 +367,8 @@ mod tests {
             ("mvn", LockfileFormat::Maven),
             ("maven", LockfileFormat::Maven),
             ("gradle", LockfileFormat::Gradle),
-            ("nuget", LockfileFormat::Msbuild),
-            ("msbuild", LockfileFormat::Msbuild),
-            ("nugetlock", LockfileFormat::NugetLock),
+            ("nuget", LockfileFormat::Nuget),
+            ("nugetlock", LockfileFormat::Nuget),
             ("go", LockfileFormat::Go),
             ("cargo", LockfileFormat::Cargo),
             ("spdx", LockfileFormat::Spdx),
@@ -403,8 +395,7 @@ mod tests {
             ("poetry", LockfileFormat::Poetry),
             ("mvn", LockfileFormat::Maven),
             ("gradle", LockfileFormat::Gradle),
-            ("msbuild", LockfileFormat::Msbuild),
-            ("nugetlock", LockfileFormat::NugetLock),
+            ("nuget", LockfileFormat::Nuget),
             ("go", LockfileFormat::Go),
             ("cargo", LockfileFormat::Cargo),
             ("spdx", LockfileFormat::Spdx),
@@ -445,14 +436,18 @@ mod tests {
             (LockfileFormat::Poetry, 2),
             (LockfileFormat::Maven, 2),
             (LockfileFormat::Gradle, 1),
-            (LockfileFormat::Msbuild, 2),
-            (LockfileFormat::NugetLock, 1),
+            (LockfileFormat::Nuget, 1),
             (LockfileFormat::Go, 1),
             (LockfileFormat::Cargo, 3),
             (LockfileFormat::Spdx, 6),
         ] {
             let mut parsed_lockfiles = Vec::new();
             for lockfile in fs::read_dir("../tests/fixtures").unwrap().flatten() {
+                // Skip directories.
+                if lockfile.file_type().unwrap().is_dir() {
+                    continue;
+                }
+
                 let lockfile_path = lockfile.path();
                 let lockfile_content = fs::read_to_string(&lockfile_path).unwrap();