+++ CHANGELOG.md
@@ -8,12 +8,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- Generating lockfiles for `*.csproj` files
+
 ### Changed
 - Include lockfile paths when analyzing projects
 - Generate and use API Keys instead of OpenID Connect tokens
 
 ### Fixed
 - Search for manifests' lockfiles in parent, rather than child directories
+
+### Removed
+- Support for `*.csproj` manifests using .NET 4 or below
 
 ## [5.5.0] - 2023-07-18